+++ changelog.md
@@ -1,3 +1,8 @@
+## Memory Game v.0.0.2
+
+-   added dummy assets
+-   create cards
+
 ## Memory Game v.0.0.1
 
 -   first commit